Create Dockerfile
Hey I was implementing point clouds in for a tool I'm working on. And This repo was one of the easier looking ones to run. Thanks for putting so much in the README. Your paper is very intereseting. Interesting how it works across a wide range of inputs.

I made a docker filer that worked well for me. Maybe others would find it useful.